新增了对管理单片机内文件夹的支持

+++ main.py
@@ -26,6 +26,7 @@ global_options = {
     "last_port"      : 0,
     "skin_mode"      : "Classic",
     "PC_PATH"        : ".\\",
+    "MCU_PATH"       :  "",
     "Now_Focus"      : "PC"
 }
 
@@ -75,7 +76,7 @@ def bind_methods():
     port_dialog.Clear.clicked.connect(port_dialog.recv_Text.clear)
     port_dialog.AutoLast.toggled.connect(func_jump_to_last_line)
     port_dialog.stopRun.clicked.connect(lambda: serial_thread.serial.write(b"\r\x03\x03"))
-    port_dialog.rebootMCU.clicked.connect(lambda: serial_thread.serial.write(b"from machine import reset\r\nreset()\r\n"))
+    port_dialog.rebootMCU.clicked.connect(lambda: serial_thread.serial.write(b"\r\x03\x03from machine import reset\r\nreset()\r\n"))
     w_p.rejected.connect(func_for_close_port_dialog)
     w_p.closeEvent = func_for_close_port_dialog
 
@@ -83,6 +84,10 @@ def bind_methods():
     shortcut_rename = QShortcut(w)
     shortcut_rename.setKey(u'F2')
     shortcut_rename.activated.connect(lambda: rename_file(main_window.PC_files.currentItem().text()))
+
+    shortcut_TAB = QShortcut(w)
+    shortcut_TAB.setKey(u'Delete')
+    shortcut_TAB.activated.connect(lambda:remove_file("PC", main_window.PC_files.currentItem().text()))
     
     shortcut_send = QShortcut(w_p)
     shortcut_send.setKey(u'Return')
@@ -90,6 +95,7 @@ def bind_methods():
     shortcut_TAB = QShortcut(w_p)
     shortcut_TAB.setKey(u'Tab')
     shortcut_TAB.activated.connect(func_for_auto_complete)
+
     labelWeb = QLabel()
     labelWeb.setText(
         '''<a style="font-family: 微软雅黑; color: #0000FF; font-size: 10pt;  text-decoration: none" href="https://github.com/umeiko/MicroPython-FileManager"> 关于本项目</a>'''
@@ -135,17 +141,27 @@ def func_for_select_port(*args):
     else:
         serial_manager.close_port()
     global_options["last_port"] = index
+    global_options["MCU_PATH"] = ""
     
 def create_right_menu_MCU():
     """右键菜单"""
-    menu   = QMenu()
-    action_1 = QAction(text="下载到本地")
-    action_1.triggered.connect(lambda:file_transport("MCU", main_window.MCU_files.currentItem().text()))
-    menu.addAction(action_1)
-    action_2 = QAction(text="删除")
-    action_2.triggered.connect(lambda:remove_file("MCU", main_window.MCU_files.currentItem().text()))
-    menu.addAction(action_2)
-    menu.exec(QCursor.pos())
+    if serial_manager.pyb is not None:
+        menu   = QMenu()
+        action_1 = QAction(text="下载到本地")
+        action_1.triggered.connect(lambda:file_transport("MCU", main_window.MCU_files.currentItem().text()))
+        menu.addAction(action_1)
+        action_2 = QAction(text="删除")
+        action_2.triggered.connect(lambda:remove_file("MCU", main_window.MCU_files.currentItem().text()))
+        menu.addAction(action_2)
+        menu.addSeparator()
+        action_4 = QAction(text="刷新")
+        action_4.triggered.connect(lambda: serial_manager.fresh_files(global_options["MCU_PATH"]))
+        menu.addAction(action_4)
+        
+        action_3 = QAction(text="新建文件夹")
+        action_3.triggered.connect(lambda:new_folder("MCU"))
+        menu.addAction(action_3)
+        menu.exec(QCursor.pos())
 
 def create_right_menu_PC():
     """右键菜单"""
@@ -154,10 +170,6 @@ def create_right_menu_PC():
     action_1.triggered.connect(lambda:file_transport("PC", main_window.PC_files.currentItem().text()))
     menu.addAction(action_1)
 
-    action_3 = QAction(text="刷新")
-    action_3.triggered.connect(fresh_PC_files)
-    menu.addAction(action_3)
-
     action_4 = QAction(text="重命名")
     action_4.triggered.connect(lambda: rename_file(main_window.PC_files.currentItem().text()))
     menu.addAction(action_4)
@@ -165,7 +177,19 @@ def create_right_menu_PC():
     action_2 = QAction(text="删除")
     action_2.triggered.connect(lambda:remove_file("PC", main_window.PC_files.currentItem().text()))
     menu.addAction(action_2)
-    menu.exec(QCursor.pos())
+
+    menu.addSeparator()
+    action_3 = QAction(text="刷新")
+    action_3.triggered.connect(fresh_PC_files)
+    menu.addAction(action_3)
+    
+    action_5 = QAction(text="新建文件夹")
+    action_5.triggered.connect(lambda: new_folder())
+    menu.addAction(action_5)
+    action_6 = QAction(text="新建代码文件")
+    action_6.triggered.connect(lambda: new_file())
+    menu.addAction(action_6)
+    menu.exec(QCursor.pos())    
 
 def func_for_send_serial_msg(*args):
     """发送串口信息的函数"""
@@ -220,20 +244,30 @@ def func_for_serial_erro(*args):
 
 def func_for_fresh_MCU_files(lst:list):
     """刷新单片机内的文件"""
+    global global_options
     main_window.MCU_files.clear()
+    if global_options["MCU_PATH"]:
+        a = QListWidgetItem()
+        a.setText("..\\")
+        icon = QIcon()
+        icon.addFile(f":/ROOT/icons/backFolder.svg", QSize(), QIcon.Normal, QIcon.Off)
+        a.setIcon(icon)
+        main_window.MCU_files.addItem(a)
     for i in lst:
         a = QListWidgetItem()
         a.setText(i.decode())
         icon = QIcon()
-        _, ext = split_file_name(i.decode())
-        icon.addFile(f":/ROOT/icons/{ext[1:]}.ico", QSize(), QIcon.Normal, QIcon.Off)
+        if i.endswith(b"/"):
+            icon.addFile(f":/ROOT/icons/folder.svg", QSize(), QIcon.Normal, QIcon.Off)
+        else:
+            _, ext = split_file_name(i.decode())
+            icon.addFile(f":/ROOT/icons/{ext[1:]}.ico", QSize(), QIcon.Normal, QIcon.Off)
         a.setIcon(icon)
         main_window.MCU_files.addItem(a)
     main_window.MCU_files.setCurrentRow(0)
 
 def fresh_PC_files():
     """刷新电脑的文件"""
-    print(global_options["PC_PATH"])
     main_window.PC_files.clear()
     files   = []
     folders = []
@@ -272,6 +306,7 @@ def fresh_PC_files():
 
 def file_transport(device:str, file_name:str, out_name:str=None):
     """传输文件 源文件设备 文件名 另存名"""
+    global global_options
     if out_name is None:
         out_name = file_name
     if serial_manager.pyb is not None:
@@ -283,20 +318,20 @@ def file_transport(device:str, file_name:str, out_name:str=None):
                 main_window.statusBar.showMessage("不支持传输该文件类型")
             try:
                 serial_manager.pyb.enter_raw_repl()
-                serial_manager.pyb.fs_put(global_options["PC_PATH"] + file_name, out_name)
-                serial_manager.fresh_files()
+                serial_manager.pyb.fs_put(global_options["PC_PATH"] + file_name, global_options["MCU_PATH"]+out_name)
+                serial_manager.fresh_files(global_options["MCU_PATH"])
                 main_window.statusBar.showMessage(f"已传输{file_name}")
             except Exception as e:
                 func_for_serial_erro(str(e))
         elif device == "MCU":
             try:
-                if os.path.exists(global_options["PC_PATH"] + file_name):
-                    main_window.statusBar.showMessage(f"本地文件{file_name}被替换")
+                if os.path.exists(global_options["PC_PATH"] + out_name):
+                    main_window.statusBar.showMessage(f"本地文件{out_name}被替换")
                 else:
-                    main_window.statusBar.showMessage(f"下载文件{file_name}到本地")
+                    main_window.statusBar.showMessage(f"下载文件{out_name}到本地")
                 serial_manager.pyb.enter_raw_repl()
-                serial_manager.pyb.fs_get(file_name, global_options["PC_PATH"] + out_name)
-                serial_manager.fresh_files()
+                serial_manager.pyb.fs_get(global_options["MCU_PATH"]+file_name, global_options["PC_PATH"] + out_name)
+                serial_manager.fresh_files(global_options["MCU_PATH"])
                 fresh_PC_files()
             except Exception as e:
                 func_for_serial_erro(str(e))
@@ -305,41 +340,59 @@ def file_transport(device:str, file_name:str, out_name:str=None):
 
 def remove_file(device:str, file_name:str):
     """删除文件"""
+    global global_options
     if device == "PC":
         if os.path.isdir(global_options["PC_PATH"] + file_name):
-            main_window.statusBar.showMessage(f"不支持删除文件夹")
+            try:
+                os.rmdir(global_options["PC_PATH"] + file_name)
+            except Exception as e:
+                main_window.statusBar.showMessage(str(e))
         else:
             os.remove(global_options["PC_PATH"] + file_name)
-            fresh_PC_files()
+        fresh_PC_files()
     elif device == "MCU":
         if serial_manager.pyb is not None:
             try:
-                serial_manager.pyb.enter_raw_repl()
-                serial_manager.pyb.fs_rm(file_name)
-                serial_manager.fresh_files()
-                main_window.statusBar.showMessage(f"已删除{file_name}")
+                if file_name.endswith('/'):
+                    serial_manager.pyb.enter_raw_repl()
+                    serial_manager.pyb.fs_rmdir(global_options["MCU_PATH"]+file_name[:-1])
+                else:
+                    serial_manager.pyb.enter_raw_repl()
+                    serial_manager.pyb.fs_rm(global_options["MCU_PATH"]+file_name)
+                serial_manager.fresh_files(global_options["MCU_PATH"])
+                main_window.statusBar.showMessage(f"已删除 \'{file_name}\'")
             except Exception as e:
-                func_for_serial_erro(str(e))
+                if len(e.args) > 1:
+                    main_window.statusBar.showMessage(f"请勿删除非空的文件夹")
+                else:
+                    func_for_serial_erro(str(e))
         else:
             main_window.statusBar.showMessage("未连接！删除无效")
 
 def open_file(device:str, file_name:str):
     """打开文件"""
+    global global_options
     if device == "PC":
         if file_name == "..\\":
-            go_pre_folder()
+            go_pre_folder("PC")
         elif os.path.isdir(global_options["PC_PATH"] + file_name):
-            open_folder(file_name)
+            open_folder(folder=file_name)
         elif file_name.endswith(supported_file_types):
             main_window.statusBar.showMessage(codeEditor.open_file(global_options["PC_PATH"]+file_name))
         else:
             main_window.statusBar.showMessage("不支持打开的文件类型")
+            fresh_PC_files()
     elif device == "MCU":
         try:
-            file_transport("MCU", file_name, "_"+file_name)
-            main_window.statusBar.showMessage(codeEditor.open_file(global_options["PC_PATH"]+"_"+file_name))
-            file_transport("PC", "_"+file_name, file_name)
-            remove_file("PC", "_"+file_name)
+            if file_name == "..\\":
+                go_pre_folder("MCU") 
+            elif not file_name.endswith("/"):
+                file_transport("MCU", file_name, "_"+file_name)
+                main_window.statusBar.showMessage(codeEditor.open_file(global_options["PC_PATH"]+"_"+file_name))
+                file_transport("PC", "_"+file_name, file_name)
+                remove_file("PC", "_"+file_name)
+            else:
+                open_folder("MCU", file_name[:-1])
         except Exception as e:
             func_for_serial_erro(str(e))
 
@@ -349,24 +402,7 @@ def rename_file(file_name):
         main_window.statusBar.showMessage(f"不支持重命名文件夹")
         return None
     fName, fType = split_file_name(file_name)
-    redialog = QDialog()    
-    shortcut = QShortcut(redialog)
-    shortcut.setKey(u'Return')
-    shortcut.activated.connect(lambda: redialog.close())
-    shortcut2 = QShortcut(redialog)
-    shortcut2.setKey(u'Enter')
-    shortcut2.activated.connect(lambda: redialog.close())
-    gridLayout = QGridLayout(redialog)
-    gridLayout.setSpacing(0)
-    gridLayout.setObjectName(u"gridLayout")
-    gridLayout.setContentsMargins(0, 0, 0, 0)
-    redialog.setWindowFlag(Qt.FramelessWindowHint)
-    redialog.move(QCursor.pos())
-    Lin = QLineEdit(redialog)
-    Lin.setText(fName)
-    gridLayout.addWidget(Lin, 0, 0, 1, 1)
-    redialog.exec()
-    after_name = Lin.text()
+    after_name = codeEditor.get_user_rename(fName)
     if not os.path.exists(global_options["PC_PATH"]+after_name+fType):
         os.rename(global_options["PC_PATH"]+file_name, global_options["PC_PATH"]+after_name+fType)
         main_window.statusBar.showMessage(f"{file_name}->{after_name+fType}")
@@ -382,22 +418,67 @@ def split_file_name(file:str):
         out += i
     return out, "."+str_list[-1]
 
-def go_pre_folder():
+def go_pre_folder(device="PC"):
     """返回上一级文件夹"""
     global global_options
-    if global_options["PC_PATH"] == ".\\":
-        return None
     nowPath = ""
-    for i in global_options["PC_PATH"].split("\\")[:-2]:
-        nowPath += i + "\\"
-    global_options["PC_PATH"] = nowPath
-    fresh_PC_files()
+    if device == "PC":
+        if global_options["PC_PATH"] == ".\\":
+            return None
+        for i in global_options["PC_PATH"].split("\\")[:-2]:
+            nowPath += i + "\\"
+        global_options["PC_PATH"] = nowPath
+        fresh_PC_files()
+    elif device == "MCU":
+        if not global_options["MCU_PATH"]:
+            return None
+        for i in global_options["MCU_PATH"].split("/")[:-2]:
+            nowPath += i + "/"
+        global_options["MCU_PATH"] = nowPath
+        serial_manager.fresh_files(global_options["MCU_PATH"])
 
-def open_folder(folder:str):
+def open_folder(device="PC", folder:str=""):
     """进入文件夹"""
     global global_options
-    global_options["PC_PATH"] += folder + "\\"
-    fresh_PC_files()
+    if device == "PC":
+        global_options["PC_PATH"] += folder + "\\"
+        fresh_PC_files()
+    if device == "MCU":
+        global_options["MCU_PATH"] += folder + "/"
+        serial_manager.fresh_files(global_options["MCU_PATH"])
+
+def new_folder(device="PC"):
+    """新建文件夹"""
+    global global_options
+    folder_name=codeEditor.get_user_rename()
+    if folder_name:
+        if device == "PC":
+            if not os.path.isdir(global_options["PC_PATH"] + folder_name):
+                os.mkdir(global_options["PC_PATH"] + '\\' + folder_name)
+            else:
+                main_window.statusBar.showMessage(f"{global_options['PC_PATH']+folder_name} already exists.")
+            fresh_PC_files()
+        elif device == "MCU":
+            try:
+                serial_manager.pyb.enter_raw_repl()
+                serial_manager.pyb.fs_mkdir(global_options["MCU_PATH"]+folder_name)
+                serial_manager.fresh_files(global_options["MCU_PATH"])
+            except Exception as e:
+                if len(e.args) > 1:
+                    main_window.statusBar.showMessage(f"尝试建立非法的文件夹")
+                else:
+                    func_for_serial_erro(str(e))
+
+def new_file():
+    """新建一个代码文件"""
+    global global_options
+    file_name=codeEditor.get_user_rename()+".py"
+    if not os.path.exists(global_options["PC_PATH"] + file_name):
+        with open(global_options["PC_PATH"] + file_name, "w", encoding="utf-8") as f:
+            f.write("# code here\n")
+        fresh_PC_files()
+    else:
+        main_window.statusBar.showMessage(f"文件 {file_name} 已存在")
 
 def main():
     bind_methods()
